Initial Progress on adding in WRE title
Renamed irtock3_religious_heads.txt to 00_a_irtock3_titles.txt so its used as a general file for converter titles. Named it in a way so that its titles are always loaded by the game, but the relevant titles can overwritten if needed. (For example, if no country is mapped to the WRE title, it will still be loaded to be used with the relevant decisions, but if a country is mapped to it and given de jure land, it properly gets overwritten so the right definition is used.

Reworked recreate_byzantine_empire_decision to properly account for new WRE title, and for other changes like allowing Roman characters to take it as well. (Will likely be modified again after I create a new culture to properly differentiate between Romanized-Greeks and Hellenistic-Greeks)